Move song loader into service class. Add Model and Controller packages.
