+++ src/main/java/spring/songs/App.java
@@ -2,13 +2,6 @@
  * This Java source file was generated by the Gradle 'init' task.
  */
 package spring.songs;
-
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.Reader;
-import org.apache.commons.csv.CSVFormat;
-import org.apache.commons.csv.CSVRecord;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -17,20 +10,5 @@ public class App {
 
     public static void main(String[] args) {
       SpringApplication.run(App.class, args);
-
-      try {
-        Reader in = new FileReader("src/main/resources/songdata.csv");
-        Iterable<CSVRecord> records = CSVFormat.EXCEL.withHeader("Artist","Song","Link","Text").parse(in);
-        for(CSVRecord record: records){
-          //just to prove the data is loading as expected and can be accessed
-          System.out.printf("Artist: %s, Song: %s, Lyrics: %s",
-              record.get("Artist"), record.get("Song"), record.get("Link"));
-        }
-      } catch (FileNotFoundException e) {
-        e.printStackTrace();
-      } catch (IOException e) {
-        e.printStackTrace();
-      }
-
     }
 }
