+++ src/main/java/spring/songs/Controller/SongController.java
@@ -1,4 +1,4 @@
-package spring.songs;
+package spring.songs.Controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
